pyproject.toml
[project]
name = "otterbridge"
version = "0.1.0"
description = "otterbridge is a lightweight, mcp server for connecting mcp client to various Large Language Model providers."
authors = [
    { name="Alex Pang", email="takyin.alex@gmail.com" },
]
license = "MIT License"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
keywords = ["llm","mcp", "mcp_server", "agent"]
url = "https://github.com/yourusername/otterbridge"
python_requires = ">3.12"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.16",
    "httpx>=0.24.1",
    "mcp[cli]>=1.6.0",
    "python-dotenv>=1.0.0",
    "websockets>=11.0.3",
]
